pyproject.toml
[tool.poetry]
name = "shield"
version = "0.2.0"
description = ""
authors = ["Jonathan Sigrist <jSigrist@web.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
grpcio-tools = "*"
pysimplegui = "*"
pyperclip = "*"
colorlog = "*"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
autopep8 = "^2.0.1"
tqdm = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"